Refactor find tests
